refactor: 게시물 테이블 분리
